+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [v0.5.0-rc] - 2018-04-10
+### Added
+- Added filter function to the mobile site in `mobile/style.css`.
+- Added feature where a marker will open its' InfoWindow if you expand the corresponding accordian item.
+
+### Fixed
+- Fixed issue where markers were not displaying the correct location information.
+
 ## [v0.4.0] - 2018-04-08
 ### Added
 - Added function, `filterEntries()` to automatically filter entries and markers based on `filterText` in `app.js`.
@@ -65,7 +73,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added preliminary files `index.html` and `app.js`.
 - Added `.gitignore` to prevent `.DS_Store` from being pushed to GitHub.
 
-[Unreleased]: https://github.com/jye0325/Neighborhood-Map/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/jye0325/Neighborhood-Map/compare/v0.5.0...HEAD
+[v0.5.0-rc]: https://github.com/jye0325/Neighborhood-Map/compare/v0.4.0...v0.5.0
+[v0.4.0]: https://github.com/jye0325/Neighborhood-Map/compare/v0.3.1...v0.4.0
 [v0.3.1]: https://github.com/jye0325/Neighborhood-Map/compare/v0.3.0...v0.3.1
 [v0.3.0]: https://github.com/jye0325/Neighborhood-Map/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/jye0325/Neighborhood-Map/compare/v0.1.0...v0.2.0